Added Average Store

+++ src/main/java/com/raghu/iot/consumer/serde/CountAndSum.java
@@ -2,9 +2,9 @@ package com.raghu.iot.consumer.serde;
 
 public class CountAndSum {
     private final Long count;
-    private final Long sum;
+    private final Double sum;
 
-    public CountAndSum(Long count, Long sum) {
+    public CountAndSum(Long count, Double sum) {
         this.count = count;
         this.sum = sum;
     }
@@ -12,7 +12,7 @@ public class CountAndSum {
     public Long getCount() {
         return count;
     }
-    public Long getSum() {
+    public Double getSum() {
         return sum;
     }
